requirements.txt
# ========================================
# RAILWAY DEPLOYMENT - VENDACERTA SYSTEM
# ========================================
# Python 3.11+ Required
# Layout Responsivo: Bootstrap 5.3.3
# Database: PostgreSQL (Production) / SQLite (Dev)

# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Compress==1.15
Flask-Caching==2.1.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Forms & Validation
WTForms==3.1.2
email-validator==2.1.0.post1

# Database Drivers
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Production WSGI Server
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0

# PDF Generation & Reports
reportlab==4.0.9
Pillow==10.1.0

# Excel Export (manylinux wheels optimized)
openpyxl==3.1.5
pandas==2.1.4
numpy==1.26.4

# Background Tasks & Scheduling
APScheduler==3.10.4

# Date & Time Utilities
pytz==2024.1
python-dateutil==2.8.2

# CLI Utilities
click==8.1.7

# Railway/Nixpacks Compatibility
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
limits==3.6.0
deprecated==1.2.14
wrapt==1.16.0
